learn/day8-12_django/day12+13/pyproject.toml
[project]
name = "day8"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.3",
    "django>=6.0.4",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "pillow>=12.2.0",
    "psycopg2-binary>=2.9.11",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "ruff>=0.14.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
